+++ example/supporting_files/steps/given_the_characters.step.dart
@@ -1,12 +1,12 @@
 import 'package:gherkin/gherkin.dart';
+
 import '../worlds/custom_world.world.dart';
 
-class GivenTheCharacters extends Given1WithWorld<String, CalculatorWorld> {
-  @override
-  Future<void> executeStep(String input1) async {
-    world.calculator.storeCharacterInput(input1);
-  }
-
-  @override
-  RegExp get pattern => RegExp(r'the characters {string}');
+StepDefinitionGeneric GivenTheCharacters() {
+  return given1<String, CalculatorWorld>(
+    'the characters {string}',
+    (input1, context) async {
+      context.world.calculator.storeCharacterInput(input1);
+    },
+  );
 }