+++ example/supporting_files/steps/given_the_numbers.step.dart
@@ -1,13 +1,12 @@
 import 'package:gherkin/gherkin.dart';
 import '../worlds/custom_world.world.dart';
 
-class GivenTheNumbers extends Given2WithWorld<num, num, CalculatorWorld> {
-  @override
-  Future<void> executeStep(num input1, num input2) async {
-    world.calculator.storeNumericInput(input1);
-    world.calculator.storeNumericInput(input2);
-  }
-
-  @override
-  RegExp get pattern => RegExp(r'the numbers {num} and {num}');
+StepDefinitionGeneric GivenTheNumbers() {
+  return given2<num, num, CalculatorWorld>(
+    'the numbers {num} and {num}',
+    (input1, input2, context) async {
+      context.world.calculator.storeNumericInput(input1);
+      context.world.calculator.storeNumericInput(input2);
+    },
+  );
 }