+++ example/supporting_files/steps/given_the_powers_of_two.step.dart
@@ -1,13 +1,12 @@
 import 'package:gherkin/gherkin.dart';
 import '../worlds/custom_world.world.dart';
 
-class GivenThePowersOfTwo extends Given2WithWorld<int, int, CalculatorWorld> {
-  @override
-  Future<void> executeStep(int input1, int input2) async {
-    world.calculator.storeNumericInput(input1);
-    world.calculator.storeNumericInput(input2);
-  }
-
-  @override
-  RegExp get pattern => RegExp(r'the powers {POW} and {POW}');
+StepDefinitionGeneric GivenThePowersOfTwo() {
+  return given2<num, num, CalculatorWorld>(
+    'the powers {POW} and {POW}',
+    (input1, input2, context) async {
+      context.world.calculator.storeNumericInput(input1);
+      context.world.calculator.storeNumericInput(input2);
+    },
+  );
 }