+++ example/supporting_files/steps/then_expect_numeric_result.step.dart
@@ -1,13 +1,12 @@
 import 'package:gherkin/gherkin.dart';
 import '../worlds/custom_world.world.dart';
 
-class ThenExpectNumericResult extends Then1WithWorld<num, CalculatorWorld> {
-  @override
-  Future<void> executeStep(num input1) async {
-    final result = world.calculator.getNumericResult();
-    expectMatch(result, input1);
-  }
-
-  @override
-  RegExp get pattern => RegExp(r'the expected result is {num}');
+StepDefinitionGeneric ThenExpectNumericResult() {
+  return given1<num, CalculatorWorld>(
+    'the expected result is {num}',
+    (input1, context) async {
+      final result = context.world.calculator.getNumericResult();
+      context.expectMatch(result, input1);
+    },
+  );
 }