+++ example/supporting_files/steps/when_numbers_are_added.step.dart
@@ -1,12 +1,11 @@
 import 'package:gherkin/gherkin.dart';
 import '../worlds/custom_world.world.dart';
 
-class WhenTheStoredNumbersAreAdded extends WhenWithWorld<CalculatorWorld> {
-  @override
-  Future<void> executeStep() async {
-    world.calculator.add();
-  }
-
-  @override
-  RegExp get pattern => RegExp(r'they are added');
+StepDefinitionGeneric WhenTheStoredNumbersAreAdded() {
+  return given<CalculatorWorld>(
+    'they are added',
+    (context) async {
+      context.world.calculator.add();
+    },
+  );
 }