+++ example/supporting_files/steps/when_the_characters_are_counted.step.dart
@@ -1,12 +1,11 @@
 import 'package:gherkin/gherkin.dart';
 import '../worlds/custom_world.world.dart';
 
-class WhenTheCharactersAreCounted extends WhenWithWorld<CalculatorWorld> {
-  @override
-  Future<void> executeStep() async {
-    world.calculator.countStringCharacters();
-  }
-
-  @override
-  RegExp get pattern => RegExp(r'they are counted');
+StepDefinitionGeneric WhenTheCharactersAreCounted() {
+  return given<CalculatorWorld>(
+    'they are counted',
+    (context) async {
+      context.world.calculator.countStringCharacters();
+    },
+  );
 }