+++ lib/src/gherkin/steps/step_definition_implementations.dart
@@ -1,3 +1,4 @@
+import '../../reporters/reporter.dart';
 import '../../expect/expect_mimic.dart';
 import './step_configuration.dart';
 import './step_definition.dart';
@@ -9,14 +10,23 @@ abstract class StepDefinitionBase<TWorld extends World>
       StepDefinitionConfiguration config, int expectParameterCount)
       : super(config, expectParameterCount);
 
-  void expect(actual, matcher, {String reason}) =>
-      ExpectMimic().expect(actual, matcher, reason: reason);
+  void expect(actual, matcher, {String reason}) => ExpectMimic().expect(
+        actual,
+        matcher,
+        reason: reason,
+      );
 
-  void expectA(actual, matcher, {String reason}) =>
-      ExpectMimic().expect(actual, matcher, reason: reason);
+  void expectA(actual, matcher, {String reason}) => ExpectMimic().expect(
+        actual,
+        matcher,
+        reason: reason,
+      );
 
-  void expectMatch(actual, matcher, {String reason}) =>
-      expect(actual, matcher, reason: reason);
+  void expectMatch(actual, matcher, {String reason}) => expect(
+        actual,
+        matcher,
+        reason: reason,
+      );
 }
 
 abstract class StepDefinition<TWorld extends World>
@@ -48,10 +58,15 @@ abstract class StepDefinition2<TWorld extends World, TInput1, TInput2>
       : super(configuration, 2);
 
   @override
-  Future<void> onRun(Iterable<dynamic> parameters) async =>
-      await executeStep(parameters.elementAt(0), parameters.elementAt(1));
+  Future<void> onRun(Iterable<dynamic> parameters) async => await executeStep(
+        parameters.elementAt(0),
+        parameters.elementAt(1),
+      );
 
-  Future<void> executeStep(TInput1 input1, TInput2 input2);
+  Future<void> executeStep(
+    TInput1 input1,
+    TInput2 input2,
+  );
 }
 
 abstract class StepDefinition3<TWorld extends World, TInput1, TInput2, TInput3>
@@ -66,7 +81,11 @@ abstract class StepDefinition3<TWorld extends World, TInput1, TInput2, TInput3>
         parameters.elementAt(2),
       );
 
-  Future<void> executeStep(TInput1 input1, TInput2 input2, TInput3 input3);
+  Future<void> executeStep(
+    TInput1 input1,
+    TInput2 input2,
+    TInput3 input3,
+  );
 }
 
 abstract class StepDefinition4<TWorld extends World, TInput1, TInput2, TInput3,
@@ -83,7 +102,11 @@ abstract class StepDefinition4<TWorld extends World, TInput1, TInput2, TInput3,
       );
 
   Future<void> executeStep(
-      TInput1 input1, TInput2 input2, TInput3 input3, TInput4 input4);
+    TInput1 input1,
+    TInput2 input2,
+    TInput3 input3,
+    TInput4 input4,
+  );
 }
 
 abstract class StepDefinition5<TWorld extends World, TInput1, TInput2, TInput3,
@@ -100,6 +123,95 @@ abstract class StepDefinition5<TWorld extends World, TInput1, TInput2, TInput3,
         parameters.elementAt(4),
       );
 
-  Future<void> executeStep(TInput1 input1, TInput2 input2, TInput3 input3,
-      TInput4 input4, TInput5 input5);
+  Future<void> executeStep(
+    TInput1 input1,
+    TInput2 input2,
+    TInput3 input3,
+    TInput4 input4,
+    TInput5 input5,
+  );
+}
+
+/// Class that contains the contextual information when a step is run
+/// Provides access to the world, reporter, step configuration and expect functions.
+class StepContext<TWorld extends World> {
+  final TWorld world;
+  final Reporter reporter;
+  final StepDefinitionConfiguration configuration;
+
+  StepContext(
+    this.world,
+    this.reporter,
+    this.configuration,
+  );
+
+  /// Assert that [actual] matches [matcher], [reason] is optional.
+  void expect(actual, matcher, {String reason}) => ExpectMimic().expect(
+        actual,
+        matcher,
+        reason: reason,
+      );
+
+  /// Assert that [actual] matches [matcher], [reason] is optional.
+  void expectA(actual, matcher, {String reason}) => expect(
+        actual,
+        matcher,
+        reason: reason,
+      );
+
+  /// Assert that [actual] matches [matcher], [reason] is optional.
+  void expectMatch(actual, matcher, {String reason}) => expect(
+        actual,
+        matcher,
+        reason: reason,
+      );
+}
+
+class GenericFunctionStepDefinition<TWorld extends World>
+    extends StepDefinitionGeneric<TWorld> {
+  final Pattern _pattern;
+  final Function _onInvoke;
+  final int _expectedParameterCount;
+
+  GenericFunctionStepDefinition(
+    this._pattern,
+    this._onInvoke,
+    this._expectedParameterCount, {
+    StepDefinitionConfiguration configuration,
+  }) : super(
+          configuration,
+          _expectedParameterCount,
+        );
+
+  @override
+  Future<void> onRun(Iterable<dynamic> parameters) {
+    final methodParams = [
+      ...parameters.take(_expectedParameterCount),
+      StepContext<TWorld>(
+        world,
+        reporter,
+        config,
+      ),
+    ];
+
+    return Function.apply(_onInvoke, methodParams);
+  }
+
+  @override
+  RegExp get pattern => _pattern is RegExp ? _pattern : RegExp(_pattern);
+}
+
+StepDefinitionGeneric<TWorld>
+    step<TWorld extends World, TInput1, TInput2, TInput3, TInput4, TInput5>(
+  Pattern pattern,
+  int expectedParameterCount,
+  Function onInvoke, {
+  StepDefinitionConfiguration configuration,
+}) {
+  return GenericFunctionStepDefinition<TWorld>(
+    pattern,
+    onInvoke,
+    expectedParameterCount,
+    configuration: configuration,
+  );
 }